api/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Matvey <motiio.none@gmail.com>"]
readme = "README.md"

[tool.pyright]
venvPath = "$HOME/Library/Caches/pypoetry/virtualenvs/"
venv = "backend-sZlJnUO7-py3.11"

[tool.mypy]
ignore_missing_imports = true
show_column_numbers = true
show_error_codes = true
pretty = true
show_error_context=true


[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.22.0"
aiogram = "^2.25.1"
asyncpg = "^0.27.0"
orjson = "^3.8.11"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.12"}
greenlet = "^2.0.2"
contextvars = "^2.4"
alembic = "^1.10.4"
python-jose = "^3.3.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.23.1"}
aioboto3 = "^11.2.0"
python-multipart = "^0.0.6"
aiofiles = "^23.1.0"
pillow = "^9.5.0"
httpx = "^0.24.1"
tenacity = "^8.2.2"
fastapi = "0.100.0-beta2"
pydantic = "^2.0"
pydantic-settings = "^2.0.0"
geoalchemy2 = "^0.14.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.265"
mypy = "^1.2.0"
factory-boy = "^3.3.0"
pytest = "^7.4.0"
freezegun = "^1.2.2"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target_version = ['py311']
include = '\.pyi?$'
ensure_newline = true

[tool.ruff]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # complexity
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "I", 'W']
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "alembic",
    "__init__.py"
]

# Same as Black.
line-length = 100

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11.
target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]
